Graphics.UI.Refluxive

Description

Refluxive framework

UI monad

data UI a

UI monad

runUI :: UI () -> IO ()

Running UI monad

mainloop :: [SomeComponent] -> UI ()

Start a mainloop, render given components as root

register :: Component UI a => ComponentView a -> UI ()

Register a ComponentView to refluxive framework

emit :: Component UI a => ComponentView a -> Signal a -> UI ()

Emit a signal from a component

watch :: (Component UI src, Component UI tgt) => ComponentView src -> (RenderState -> Signal src -> StateT (Model tgt) UI ()) -> Watcher UI tgt

Watcher function

asRoot :: Component UI a => ComponentView a -> SomeComponent

A function which is used for passing to mainloop

setClearColor :: Color -> UI ()

Set background color

addWatchSignal :: Component UI tgt => ComponentView tgt -> Watcher UI tgt -> UI ()

Add a watch signal

quit :: UI ()

Quit the mainloop

Component

Raw Component

Raw component is a component of static picture with no signal

Component UI "raw" has Graphical value as model

rawGraphical :: ComponentView "raw" -> Graphical -> UI ()

Register a graphical object to raw component | This will override the previous graphical

Built-in Component

Built-in component is a component which emits SDL events as BuiltInSignal

_builtIn :: Lens' UIState (ComponentView "builtin")

A lens to focus on builtIn component

Component types

class KnownSymbol a => Component m a | a -> m where

Component type

Minimal complete definition

newModel, getGraphical

Associated Types

type ModelParam a

Parameter, passing to the constructor

data Model a

Model datatype

data Signal a

Signal datatype

Methods

newModel :: MonadIO m => ModelParam a -> m (Model a)

Creating new model from ModelParam, it is called internally in new function

initComponent :: MonadIO m => ComponentView a -> m ()

Initialize function, it is called internally after new function

getGraphical :: MonadIO m => Model a -> m Graphical

Projecting model to view

data ComponentView a

An instance for Component

data Watcher m tgt

Watcher type with context m and target ui name tgt

getModel :: MonadIO m => ComponentView a -> m (Model a)

Get current model

Operations on component

new :: Component UI a => ModelParam a -> UI (ComponentView a)

Constructor of a component

view :: Component UI a => ComponentView a -> UI Graphical

Get the view with current snapshot of model

operateModel :: Component UI a => ComponentView a -> StateT (Model a) UI r -> UI r

Modify internal model of a component

Re-exports

module Graphics.UI.Refluxive.Graphical

module Data.Default

Orphan instances

Component UI "builtin"

Instance details

Associated Types

type ModelParam "builtin" :: *

data Model "builtin" :: *

data Signal "builtin" :: *

Methods

newModel :: ModelParam "builtin" -> UI (Model "builtin")

initComponent :: ComponentView "builtin" -> UI ()

getGraphical :: Model "builtin" -> UI Graphical

Component UI "raw"

Instance details

Associated Types

type ModelParam "raw" :: *

data Model "raw" :: *

data Signal "raw" :: *

Methods

newModel :: ModelParam "raw" -> UI (Model "raw")

initComponent :: ComponentView "raw" -> UI ()

getGraphical :: Model "raw" -> UI Graphical