Graphics.UI.Refluxive.Graphical

Description

Provides graphical functions

Graphical types

data Graphical

Graphical object can be displayed on screen with render

data RenderState

A rendering context

Constructors

RenderState

Fields

color :: Color

coordinate :: Pos

scaler :: Pos

defRenderState :: RenderState

Default rendering state

render

Arguments

:: MonadIO m

=> Color
clearColor (not a rendering color)

-> Maybe Font
font data

-> Renderer
current renderer

-> Graphical
object to render

-> (RenderState -> String -> m ())
tagging function which is used in Components

-> m ()

A function to draw Graphical objects

Drawing functions

empty :: Graphical

Empty object

lineTo :: Pos -> Pos -> Graphical

Line object

relLineTo :: Pos -> Pos -> Graphical

Line object, as target position is calculated relatively

text :: Text -> Graphical

Text object (font family and size is currently hard-coded)

textWith :: IncludeAssoc TextStyle xs => Record xs -> Text -> Graphical

Text object with text style

type TextStyle = '["styles" >: [Style]]

Text style for text object

gridLayout :: Pos -> Graphical -> Graphical

Scaling function, useful to place objects in grid

rectangle :: Pos -> Pos -> Graphical

Rectangle object

rectangle == rectangleWith nil

rectangleWith

Arguments

:: IncludeAssoc ShapeStyle xs

=> Record xs
shape style

-> Pos
top-left coordinate

-> Pos
size

-> Graphical

Rectangle object with shape style

type ShapeStyle = ["fill" >: Bool, "rounded" >: Maybe Int]

Shape style for rectangle object

data Style :: *

Possible styles that can be applied to a Font.

Constructors

Bold

Italic

Underline

Strikethrough

colored :: Color -> Graphical -> Graphical

Coloring function

translate :: Pos -> Graphical -> Graphical

Coordinate translating function

graphics :: [Graphical] -> Graphical

Many graphical objects

clip :: Pos -> Graphical -> Graphical

Clipping function

viewInfo :: String -> Graphical -> Graphical

Tagging id with given graphical (like CSS id)