Graphics.UI.Refluxive

Description

Refluxive framework

UI monad

data UI a

UI monad, where many operations can perform

runUI :: UI () -> IO ()

Running UI monad

mainloop :: [SomeComponent] -> UI ()

Start a mainloop, render given components as root

register :: Component UI a => ComponentView a -> UI ()

Register a ComponentView to refluxive framework

emit :: Component UI a => ComponentView a -> Signal a -> UI ()

Emit a signal from a component

watch :: (Component UI src, Component UI tgt) => ComponentView src -> (RenderState -> Signal src -> StateT (Model tgt) UI ()) -> Watcher UI tgt

Watcher function

asRoot :: Component UI a => ComponentView a -> SomeComponent

A function which is used for passing to mainloop

setClearColor :: Color -> UI ()

Set background color

addWatchSignal :: Component UI tgt => ComponentView tgt -> Watcher UI tgt -> UI ()

Add a watch signal

quit :: UI ()

Quit the mainloop

Component "raw"

rawGraphical :: ComponentView "raw" -> Graphical -> ComponentView "raw"

Register a graphical object to raw component

Re-exports

module Graphics.UI.Refluxive.Graphical

Component types

class KnownSymbol a => Component m a | a -> m where

Component type

Minimal complete definition

newModel, getGraphical

Associated Types

type ModelParam a

Parameter, passing to the constructor

data Model a

Model datatype

data Signal a

Signal datatype

Methods

uid :: proxy a -> String

newModel :: MonadIO m => ModelParam a -> m (Model a)

Creating new model from ModelParam, it is called internally in new function

initComponent :: MonadIO m => ComponentView a -> m ()

Initialize function, it is called internally after new function

getGraphical :: MonadIO m => Model a -> m Graphical

Projecting model to view

data ComponentView a

An instance for Component

data Watcher m tgt

Watcher type with context m and target ui name tgt

getModel :: ComponentView a -> Model a

Get current model

Signal

_builtIn :: Lens' UIState (ComponentView "builtin")

A lens to focus on builtIn component

Operations on component

fromModel :: Component UI a => Model a -> UI (ComponentView a)

new :: Component UI a => ModelParam a -> UI (ComponentView a)

Constructor of a component

view :: Component UI a => ComponentView a -> UI Graphical

Get the view with current snapshot of model

operateModel :: Component UI a => ComponentView a -> StateT (Model a) UI () -> UI ()

Modify internal model of a component

Orphan instances

Component UI "builtin"

Associated Types

type ModelParam ("builtin" :: Symbol) :: *

data Model ("builtin" :: Symbol) :: *

data Signal ("builtin" :: Symbol) :: *

Methods

uid :: proxy "builtin" -> String

newModel :: ModelParam "builtin" -> UI (Model "builtin")

initComponent :: ComponentView "builtin" -> UI ()

getGraphical :: Model "builtin" -> UI Graphical

Component UI "raw"

Associated Types

type ModelParam ("raw" :: Symbol) :: *

data Model ("raw" :: Symbol) :: *

data Signal ("raw" :: Symbol) :: *

Methods

uid :: proxy "raw" -> String

newModel :: ModelParam "raw" -> UI (Model "raw")

initComponent :: ComponentView "raw" -> UI ()

getGraphical :: Model "raw" -> UI Graphical