Graphics.UI.Refluxive.Component

Description

UI Components

Documentation

class KnownSymbol a => Component m a | a -> m where

Component type

Minimal complete definition

newModel, getGraphical

Associated Types

type ModelParam a

Parameter, passing to the constructor

data Model a

Model datatype

data Signal a

Signal datatype

Methods

newModel :: MonadIO m => ModelParam a -> m (Model a)

Creating new model from ModelParam, it is called internally in new function

initComponent :: MonadIO m => ComponentView a -> m ()

Initialize function, it is called internally after new function

getGraphical :: MonadIO m => Model a -> m Graphical

Projecting model to view

data ComponentView a

An instance for Component

Constructors

ComponentView

Fields

modelRef :: IORef (Model a)

name :: String

renderStateRef :: IORef RenderState

data Watcher m tgt

Watcher type with context m and target ui name tgt

Constructors

(Component m src, Component m tgt) => Watcher (ComponentView src) (RenderState -> Signal src -> StateT (Model tgt) m ())

getModel :: MonadIO m => ComponentView a -> m (Model a)

Get current model

prefix :: Component m a => proxy a -> String